chore: refactor GlobalSettings and SwapLiquiditySettings
